refactor: convert models.py to module structure organized by swagger tags
- Create ABConnect/api/models/ module with separate files for each tag
- base.py: Base models, enums, and utility functions
- addresses.py: All address-related models
- companies.py: Company models with inheritance hierarchy
- contacts.py: Contact model
- jobs.py: Job and related models (JobItem, JobTask, etc.)
- users.py: User model
- documents.py: Document model
- forms.py: Form and FormField models
- lookups.py: Lookup enums and values
- responses.py: Response wrapper models
- Delete models_backup.py as it's in version control
- Maintain backward compatibility with comprehensive __init__.py exports
