--- ABConnect/api/models.py
@@ -1,799 +1,0 @@
-"""Response models for API data structures.
-
-This module provides data models and type definitions for API responses,
-organized by swagger tags for better structure and maintainability.
-"""
-
-from datetime import datetime
-from typing import Any, Dict, List, Optional, Union
-from enum import Enum
-from pydantic import BaseModel, Field, ConfigDict
-
-
-# ==============================================================================
-# BASE MODELS AND ENUMS
-# ==============================================================================
-
-class ABConnectBaseModel(BaseModel):
-    """Base class for all API models."""
-
-    model_config = ConfigDict(extra="allow", populate_by_name=True)
-
-    id: Optional[str] = None
-    created: Optional[datetime] = None
-    modified: Optional[datetime] = None
-
-
-# ==============================================================================
-# COMMON ENUMS
-# ==============================================================================
-
-class CompanyType(str, Enum):
-    """Valid company types from ABC API."""
-    AGENT = "Agent"
-    CARRIER = "Carrier"
-    CORPORATE = "Corporate"
-    CUSTOMER = "Customer"
-    FRANCHISEE = "Franchisee"
-    NATIONAL_ACCOUNT = "National Account"
-    TERMINAL = "Terminal"
-    VENDOR = "Vendor"
-
-
-# Company type ID mapping for reference
-COMPANY_TYPE_IDS = {
-    "Agent": "697cc861-d271-4baf-8cbb-2eb055a1005a",
-    "Carrier": "88a541e1-456e-4e6e-b445-af75311b694f",
-    "Corporate": "8ec06e36-7e6a-4ed6-a27c-7cc0c13a7292",
-    "Customer": "8e809044-8d69-4618-9533-265d7e71db13",
-    "Franchisee": "e7f85166-34cf-429b-805d-261b44cb0c04",
-    "National Account": "27654fb3-9507-e811-8f3f-00155d426802",
-    "Terminal": "65d232c9-3031-4682-83b5-594da868d9dd",
-    "Vendor": "4176c2d7-b7ae-ec11-822e-a4aa13c701a3",
-}
-
-
-class JobStatus(str, Enum):
-    """Job status enumeration."""
-    DRAFT = "draft"
-    PENDING = "pending"
-    ACTIVE = "active"
-    IN_PROGRESS = "in_progress"
-    COMPLETED = "completed"
-    CANCELLED = "cancelled"
-    FAILED = "failed"
-
-
-class TaskStatus(str, Enum):
-    """Task status enumeration."""
-    PENDING = "pending"
-    SCHEDULED = "scheduled"
-    IN_PROGRESS = "in_progress"
-    COMPLETED = "completed"
-    CANCELLED = "cancelled"
-
-
-class ContactType(str, Enum):
-    """Contact type enumeration."""
-    PRIMARY = "primary"
-    BILLING = "billing"
-    SHIPPING = "shipping"
-    TECHNICAL = "technical"
-    OTHER = "other"
-
-
-# ==============================================================================
-# ADDRESS MODELS
-# ==============================================================================
-
-class AddressCoordinates(BaseModel):
-    """Coordinate model for addresses."""
-    latitude: Optional[float] = None
-    longitude: Optional[float] = None
-
-
-class AddressBase(BaseModel):
-    """Base address model with common fields."""
-    address1: Optional[str] = None
-    address2: Optional[str] = None
-    city: Optional[str] = None
-    state: Optional[str] = None
-    zipCode: Optional[str] = None
-    countryCode: Optional[str] = None
-    countryName: Optional[str] = None
-
-
-class Address(AddressBase, ABConnectBaseModel):
-    """Standard address model."""
-    line1: Optional[str] = None  # Alias for address1
-    line2: Optional[str] = None  # Alias for address2
-    line3: Optional[str] = None
-    zip: Optional[str] = None  # Alias for zipCode
-    country: Optional[str] = None  # Alias for countryCode
-    type: Optional[str] = None
-    isValid: Optional[bool] = None
-
-
-class MainAddress(AddressBase):
-    """Main address model with extended fields."""
-    id: Optional[int] = None
-    isValid: Optional[bool] = None
-    dontValidate: Optional[bool] = None
-    propertyType: Optional[str] = None
-    address1Value: Optional[str] = None
-    address2Value: Optional[str] = None
-    countryId: Optional[str] = None
-    latitude: Optional[float] = None
-    longitude: Optional[float] = None
-    fullCityLine: Optional[str] = None
-    coordinates: Optional[AddressCoordinates] = None
-
-
-class AddressData(AddressBase):
-    """Address data model for company details."""
-    company: Optional[str] = None
-    firstLastName: Optional[str] = None
-    addressLine1: Optional[str] = None
-    addressLine2: Optional[str] = None
-    contactBOLNote: Optional[str] = None
-    stateCode: Optional[str] = None
-    propertyType: Optional[str] = None
-    fullCityLine: Optional[str] = None
-    phone: Optional[str] = None
-    cellPhone: Optional[str] = None
-    fax: Optional[str] = None
-    email: Optional[str] = None
-    addressLine2Visible: Optional[bool] = None
-    companyVisible: Optional[bool] = None
-    countryNameVisible: Optional[bool] = None
-    phoneVisible: Optional[bool] = None
-    emailVisible: Optional[bool] = None
-    fullAddressLine: Optional[str] = None
-    fullAddress: Optional[str] = None
-    countryId: Optional[str] = None
-
-
-# ==============================================================================
-# COMPANIES MODELS
-# ==============================================================================
-
-class CompanyBase(ABConnectBaseModel):
-    """Base company model with fields common to all company responses."""
-    code: Optional[str] = None
-    name: Optional[str] = None
-    parentCompanyId: Optional[str] = None
-    
-    # Common fields that appear in multiple responses
-    companyTypeId: Optional[str] = Field(None, alias="companyTypeID")
-    companyEmail: Optional[str] = None
-    companyPhone: Optional[str] = None
-    isActive: Optional[bool] = None
-
-
-class CompanyBasic(CompanyBase):
-    """Basic company model returned by /api/companies/{id}.
-    
-    This is the simplest company response with just the essential fields.
-    """
-    pass  # All fields inherited from CompanyBase
-
-
-class CompanyInfo(CompanyBase):
-    """Company info model used within other responses."""
-    companyId: Optional[str] = None
-    companyDisplayId: Optional[str] = None
-    companyName: Optional[str] = None
-    companyCode: Optional[str] = None
-    thumbnailLogo: Optional[str] = None
-    companyLogo: Optional[str] = None
-    mapsMarkerImage: Optional[str] = None
-    mainAddress: Optional[MainAddress] = None
-    isThirdParty: Optional[bool] = None
-    isHidden: Optional[bool] = None
-
-
-class OverridableValue(BaseModel):
-    """Model for overridable values in company details."""
-    defaultValue: Optional[str] = None
-    overrideValue: Optional[str] = None
-    forceEmpty: Optional[bool] = None
-    value: Optional[str] = None
-
-
-class OverridableAddressData(BaseModel):
-    """Overridable address data model."""
-    company: Optional[OverridableValue] = None
-    firstLastName: Optional[OverridableValue] = None
-    addressLine1: Optional[OverridableValue] = None
-    addressLine2: Optional[OverridableValue] = None
-    city: Optional[OverridableValue] = None
-    state: Optional[OverridableValue] = None
-    zipCode: Optional[OverridableValue] = None
-    phone: Optional[OverridableValue] = None
-    email: Optional[OverridableValue] = None
-    fullAddressLine: Optional[str] = None
-    fullAddress: Optional[OverridableValue] = None
-    fullCityLine: Optional[OverridableValue] = None
-
-
-class CompanyDetails(CompanyBase):
-    """Detailed company model returned by /api/companies/{companyId}/details.
-    
-    This includes all fields from CompanyBase plus extensive additional details.
-    """
-    # User and contact information
-    userId: Optional[str] = Field(None, alias="userID")
-    contactName: Optional[str] = None
-    contactPhone: Optional[str] = None
-    primaryContactId: Optional[int] = None
-    payerContactId: Optional[int] = None
-    payerContactName: Optional[str] = None
-    primaryCustomerName: Optional[str] = None
-    contactInfo: Optional[str] = None
-    
-    # Company type and classification
-    companyType: Optional[str] = None
-    parentCompanyName: Optional[str] = None
-    parentCompanyID: Optional[str] = None
-    franchiseeName: Optional[str] = None
-    
-    # Contact details
-    companyFax: Optional[str] = None
-    companyWebSite: Optional[str] = None
-    companyCell: Optional[str] = None
-    customerCell: Optional[str] = None
-    
-    # Business information
-    industryType: Optional[str] = None
-    industryTypeName: Optional[str] = None
-    taxId: Optional[str] = None
-    companyDisplayID: Optional[str] = None
-    
-    # Logos and branding
-    companyLogo: Optional[str] = None
-    letterHeadLogo: Optional[str] = None
-    thumbnailLogo: Optional[str] = None
-    mapsMarkerImage: Optional[str] = None
-    colorTheme: Optional[str] = None
-    
-    # Settings and preferences
-    parcelOnly: Optional[bool] = None
-    isThirdParty: Optional[bool] = None
-    isGlobal: Optional[bool] = None
-    isQbUser: Optional[bool] = None
-    skipIntacct: Optional[bool] = None
-    isAccess: Optional[bool] = None
-    isPrefered: Optional[bool] = None
-    isHide: Optional[bool] = None
-    isDontUse: Optional[bool] = None
-    
-    # Codes and references
-    pzCode: Optional[str] = None
-    referralCode: Optional[str] = None
-    
-    # Pricing and insurance
-    franchiseeMaturityType: Optional[str] = None
-    pricingToUse: Optional[str] = None
-    insuranceType: Optional[str] = None
-    wholeSaleMarkup: Optional[float] = None
-    baseMarkup: Optional[float] = None
-    mediumMarkup: Optional[float] = None
-    highMarkup: Optional[float] = None
-    companyInsurancePricing: Optional[float] = None
-    companyServicePricing: Optional[float] = None
-    companyTaxPricing: Optional[float] = None
-    
-    # Carrier and account management
-    accountManagerFranchiseeId: Optional[str] = None
-    accountManagerFranchiseeName: Optional[str] = None
-    carrierAccountsSourceCompanyId: Optional[str] = None
-    carrierAccountsSourceCompanyName: Optional[str] = None
-    
-    # API settings
-    autoPriceAPIEnableEmails: Optional[bool] = None
-    autoPriceAPIEnableSMSs: Optional[bool] = None
-    
-    # Capabilities and statistics
-    commercialCapabilities: Optional[int] = None
-    totalJobs: Optional[int] = None
-    totalJobsRevenue: Optional[float] = None
-    totalSales: Optional[int] = None
-    totalSalesRevenue: Optional[float] = None
-    
-    # Address information
-    mainAddress: Optional[MainAddress] = None
-    addressData: Optional[AddressData] = None
-    overridableAddressData: Optional[OverridableAddressData] = None
-    companyInfo: Optional[CompanyInfo] = None
-    
-    # Legacy fields
-    companyID: Optional[str] = None
-    addressID: Optional[int] = None
-    address: Optional[str] = None
-    address1: Optional[str] = None
-    address2: Optional[str] = None
-    city: Optional[str] = None
-    state: Optional[str] = None
-    stateCode: Optional[str] = None
-    countryName: Optional[str] = None
-    countryCode: Optional[str] = None
-    countryID: Optional[str] = None
-    zipCode: Optional[str] = None
-    latitude: Optional[float] = None
-    longitude: Optional[float] = None
-    
-    # Metadata
-    totalRows: Optional[int] = None
-    depth: Optional[int] = None
-    miles: Optional[float] = None
-    createdUser: Optional[str] = None
-    createdDate: Optional[datetime] = None
-    createdBy: Optional[str] = None
-    modifiedDate: Optional[datetime] = None
-    modifiedBy: Optional[str] = None
-    
-    # Other fields
-    result: Optional[str] = None
-    addressMappingID: Optional[int] = None
-    contactID: Optional[int] = None
-    mappingLocations: Optional[str] = None
-    locationCount: Optional[int] = None
-    baseParent: Optional[str] = None
-    copyMaterialFrom: Optional[str] = None
-
-
-# Company Full Details Models (for /api/companies/{companyId}/fulldetails)
-
-class FileUpload(BaseModel):
-    """File upload model for logos."""
-    filePath: Optional[str] = None
-    newFile: Optional[str] = None
-
-
-class CompanyDetailsSection(BaseModel):
-    """Details section of company full details."""
-    displayId: Optional[str] = None
-    name: Optional[str] = None
-    taxId: Optional[str] = None
-    code: Optional[str] = None
-    parentId: Optional[str] = None
-    franchiseeId: Optional[str] = None
-    companyTypeId: Optional[str] = None
-    industryTypeId: Optional[str] = None
-    cellPhone: Optional[str] = None
-    phone: Optional[str] = None
-    fax: Optional[str] = None
-    email: Optional[str] = None
-    website: Optional[str] = None
-    isActive: Optional[bool] = None
-    isHidden: Optional[bool] = None
-    isGlobal: Optional[bool] = None
-    isNotUsed: Optional[bool] = None
-    isPreferred: Optional[bool] = None
-    payerContactId: Optional[int] = None
-    payerContactName: Optional[str] = None
-
-
-class CompanyPreferences(BaseModel):
-    """Company preferences model."""
-    companyHeaderLogo: Optional[FileUpload] = None
-    thumbnailLogo: Optional[FileUpload] = None
-    letterHeadLogo: Optional[FileUpload] = None
-    mapsMarker: Optional[FileUpload] = None
-    isQbUser: Optional[bool] = None
-    skipIntacct: Optional[bool] = None
-    pricingToUse: Optional[str] = None
-    pzCode: Optional[str] = None
-    insuranceTypeId: Optional[str] = None
-    franchiseeMaturityTypeId: Optional[str] = None
-    isCompanyUsedAsCarrierSource: Optional[bool] = None
-    carrierAccountsSourceCompanyId: Optional[str] = None
-    carrierAccountsSourceCompanyName: Optional[str] = None
-    accountManagerFranchiseeId: Optional[str] = None
-    accountManagerFranchiseeName: Optional[str] = None
-    autoPriceAPIEnableEmails: Optional[bool] = None
-    autoPriceAPIEnableSMSs: Optional[bool] = None
-    copyMaterials: Optional[int] = None
-
-
-class TransportationCharge(BaseModel):
-    """Transportation charge model."""
-    baseTripFee: Optional[float] = None
-    baseTripMile: Optional[float] = None
-    extraFee: Optional[float] = None
-    fuelSurcharge: Optional[float] = None
-
-
-class MarkupRates(BaseModel):
-    """Markup rates model."""
-    wholeSale: Optional[float] = None
-    base: Optional[float] = None
-    medium: Optional[float] = None
-    high: Optional[float] = None
-
-
-class CompanyPricing(BaseModel):
-    """Company pricing model."""
-    transportationCharge: Optional[TransportationCharge] = None
-    transportationMarkups: Optional[MarkupRates] = None
-    carrierFreightMarkups: Optional[MarkupRates] = None
-
-
-class InsuranceOption(BaseModel):
-    """Insurance option model."""
-    insuranceSlabId: Optional[str] = None
-    option: Optional[int] = None
-    sellPrice: Optional[float] = None
-
-
-class CompanyInsurance(BaseModel):
-    """Company insurance model."""
-    isp: Optional[InsuranceOption] = None
-    nsp: Optional[InsuranceOption] = None
-    ltl: Optional[InsuranceOption] = None
-
-
-class CompanyFullDetails(CompanyBase):
-    """Full company details returned by /api/companies/{companyId}/fulldetails.
-    
-    This inherits from CompanyBase but overrides the ID field and adds sections.
-    """
-    id: Optional[str] = None  # This endpoint returns id at root level
-    details: Optional[CompanyDetailsSection] = None
-    preferences: Optional[CompanyPreferences] = None
-    capabilities: Optional[int] = None
-    address: Optional[MainAddress] = None
-    pricing: Optional[CompanyPricing] = None
-    insurance: Optional[CompanyInsurance] = None
-    readOnlyAccess: Optional[bool] = None
-
-
-# ==============================================================================
-# CONTACTS MODELS
-# ==============================================================================
-
-class Contact(ABConnectBaseModel):
-    """Contact model."""
-    firstName: Optional[str] = None
-    lastName: Optional[str] = None
-    email: Optional[str] = None
-    phone: Optional[str] = None
-    mobile: Optional[str] = None
-    title: Optional[str] = None
-    department: Optional[str] = None
-    companyId: Optional[str] = None
-    contactType: Optional[ContactType] = None
-    isActive: bool = True
-
-
-# ==============================================================================
-# JOB MODELS
-# ==============================================================================
-
-class JobItem(ABConnectBaseModel):
-    """Item within a job."""
-    code: Optional[str] = None
-    name: Optional[str] = None
-    description: Optional[str] = None
-    type: Optional[str] = None
-    weight: Optional[float] = None
-    dimensions: Optional[Dict[str, float]] = None
-    value: Optional[float] = None
-    quantity: Optional[int] = None
-    currency: str = "USD"
-    isHazmat: bool = False
-    metadata: Dict[str, Any] = Field(default_factory=dict)
-
-
-class JobTask(ABConnectBaseModel):
-    """Task within a job."""
-    jobId: Optional[str] = None
-    type: Optional[str] = None
-    status: Optional[TaskStatus] = None
-    description: Optional[str] = None
-    assignedTo: Optional[str] = None
-    scheduledDate: Optional[datetime] = None
-    completedDate: Optional[datetime] = None
-    duration: Optional[int] = None
-    notes: Optional[str] = None
-    metadata: Dict[str, Any] = Field(default_factory=dict)
-
-
-class JobShipment(ABConnectBaseModel):
-    """Shipment information within a job."""
-    trackingNumber: Optional[str] = None
-    carrier: Optional[str] = None
-    service: Optional[str] = None
-    estimatedDelivery: Optional[datetime] = None
-    actualDelivery: Optional[datetime] = None
-    status: Optional[str] = None
-    weight: Optional[float] = None
-    cost: Optional[float] = None
-
-
-class JobPayment(ABConnectBaseModel):
-    """Payment information within a job."""
-    method: Optional[str] = None
-    amount: Optional[float] = None
-    currency: str = "USD"
-    status: Optional[str] = None
-    transactionId: Optional[str] = None
-    processedDate: Optional[datetime] = None
-
-
-class Job(ABConnectBaseModel):
-    """Job model with nested entities."""
-    displayId: Optional[str] = None
-    code: Optional[str] = None
-    type: Optional[str] = None
-    status: Optional[JobStatus] = None
-    customerId: Optional[str] = None
-    customerName: Optional[str] = None
-    vendorId: Optional[str] = None
-    vendorName: Optional[str] = None
-    
-    # Addresses
-    originAddress: Optional[Address] = None
-    destinationAddress: Optional[Address] = None
-    
-    # Nested entities
-    items: List[JobItem] = Field(default_factory=list)
-    tasks: List[JobTask] = Field(default_factory=list)
-    shipments: List[JobShipment] = Field(default_factory=list)
-    payments: List[JobPayment] = Field(default_factory=list)
-    
-    # Dates
-    scheduledDate: Optional[datetime] = None
-    completedDate: Optional[datetime] = None
-    
-    # Totals
-    totalWeight: Optional[float] = None
-    totalValue: Optional[float] = None
-    totalCost: Optional[float] = None
-    
-    # Additional fields
-    notes: Optional[str] = None
-    metadata: Dict[str, Any] = Field(default_factory=dict)
-
-
-# ==============================================================================
-# USERS MODELS
-# ==============================================================================
-
-class User(ABConnectBaseModel):
-    """User model."""
-    username: Optional[str] = None
-    email: Optional[str] = None
-    firstName: Optional[str] = None
-    lastName: Optional[str] = None
-    role: Optional[str] = None
-    permissions: List[str] = Field(default_factory=list)
-    companyId: Optional[str] = None
-    isActive: bool = True
-    lastLogin: Optional[datetime] = None
-
-
-# ==============================================================================
-# DOCUMENTS MODELS
-# ==============================================================================
-
-class Document(ABConnectBaseModel):
-    """Document model."""
-    name: Optional[str] = None
-    type: Optional[str] = None
-    size: Optional[int] = None
-    mimeType: Optional[str] = None
-    url: Optional[str] = None
-    jobId: Optional[str] = None
-    uploadedBy: Optional[str] = None
-    tags: List[str] = Field(default_factory=list)
-    metadata: Dict[str, Any] = Field(default_factory=dict)
-
-
-# ==============================================================================
-# FORMS MODELS
-# ==============================================================================
-
-class FormField(BaseModel):
-    """Form field definition."""
-    name: str
-    type: str
-    label: Optional[str] = None
-    required: bool = False
-    defaultValue: Optional[Any] = None
-    options: Optional[List[str]] = None
-    validation: Optional[Dict[str, Any]] = None
-
-
-class Form(ABConnectBaseModel):
-    """Form/Template model."""
-    name: Optional[str] = None
-    type: Optional[str] = None
-    version: Optional[str] = None
-    fields: List[FormField] = Field(default_factory=list)
-    isActive: bool = True
-    metadata: Dict[str, Any] = Field(default_factory=dict)
-
-
-# ==============================================================================
-# LOOKUP MODELS
-# ==============================================================================
-
-class LookupKeys(str, Enum):
-    """Available lookup keys."""
-    JOBINTACCTSTATUS = " JobIntacctStatus"
-    BASISTYPES = "BasisTypes"
-    CANCELLEDTYPES = "CancelledTypes"
-    CFILLTYPE = "CFillType"
-    COMMODITYCATEGORY = "CommodityCategory"
-    COMPANYTYPES = "CompanyTypes"
-    CONTACTTYPES = "ContactTypes"
-    CONTAINERTYPE = "ContainerType"
-    CPACKTYPE = "CPackType"
-    CREDITCARDTYPES = "CreditCardTypes"
-    DOCUMENTTAGS = "DocumentTags"
-    FOLLOWUPHEATOPTION = "FollowupHeatOption"
-    FOLLOWUPPIPELINEOPTION = "FollowupPipelineOption"
-    FRANCHISEETYPES = "FranchiseeTypes"
-    FREIGHTCLASS = "FreightClass"
-    FREIGHTTYPES = "FreightTypes"
-    INDUSTRYTYPES = "IndustryTypes"
-    INSURANCEOPTION = "InsuranceOption"
-    INSURANCETYPE = "InsuranceType"
-    ITEMNOTEDCONDITIONS = "ItemNotedConditions"
-    ITEMTYPES = "ItemTypes"
-    JOBMANAGEMENT = "Job Management Status"
-    JOBMGMTTYPES = "JobMgmtTypes"
-    JOBNOTECATEGORY = "JobNoteCategory"
-    JOBSSTATUSTYPES = "JobsStatusTypes"
-    JOBTYPE = "JobType"
-    ONHOLDNEXTSTEP = "OnHoldNextStep"
-    ONHOLDREASON = "OnHoldReason"
-    ONHOLDRECOLVEDCODE = "OnHoldRecolvedCode"
-    PAYMENTSTATUSES = "PaymentStatuses"
-    PRICINGTOUSE = "PricingToUse"
-    QBJOBTRANSTYPE = "QBJobTransType"
-    QBWSTRANSTYPE = "QBWSTransType"
-    RESPONSIBILITYPARTY = "ResponsibilityParty"
-    ROOMTYPES = "RoomTypes"
-    TRANSRULES = "TransRules"
-    TRANSTYPES = "TransTypes"
-    YESNO = "YesNo"
-
-
-class LookupValue(BaseModel):
-    """Lookup value model."""
-    id: str
-    value: str
-    description: Optional[str] = None
-    isActive: bool = True
-    sortOrder: Optional[int] = None
-    metadata: Optional[Dict[str, Any]] = None
-
-
-# ==============================================================================
-# RESPONSE WRAPPER MODELS
-# ==============================================================================
-
-class PaginatedResponse(BaseModel):
-    """Paginated response wrapper."""
-    
-    model_config = ConfigDict(extra="allow")
-    
-    data: List[Any] = Field(default_factory=list)
-    page: int = 1
-    per_page: int = 50
-    total: int = 0
-    total_pages: int = 0
-    has_next: bool = False
-    has_prev: bool = False
-    
-    @classmethod
-    def from_dict(cls, data: Dict[str, Any], item_class: Optional[type] = None) -> "PaginatedResponse":
-        """Create paginated response from dictionary."""
-        page = data.get("page", 1)
-        per_page = data.get("per_page", data.get("perPage", 50))
-        total = data.get("total", 0)
-        total_pages = data.get("total_pages", data.get("totalPages", 0))
-        has_next = data.get("has_next", data.get("hasNext", False))
-        has_prev = data.get("has_prev", data.get("hasPrev", False))
-        
-        items = data.get("data", data.get("items", data.get("results", [])))
-        
-        if item_class and hasattr(item_class, "model_validate"):
-            parsed_items = [item_class.model_validate(item) for item in items]
-        else:
-            parsed_items = items
-        
-        return cls(
-            data=parsed_items,
-            page=page,
-            per_page=per_page,
-            total=total,
-            total_pages=total_pages,
-            has_next=has_next,
-            has_prev=has_prev,
-        )
-
-
-class ErrorResponse(BaseModel):
-    """Error response model."""
-    error: str
-    message: str
-    code: Optional[str] = None
-    details: Optional[Dict[str, Any]] = None
-    timestamp: Optional[datetime] = Field(default_factory=datetime.now)
-
-
-# ==============================================================================
-# MODEL REGISTRY
-# ==============================================================================
-
-MODEL_REGISTRY = {
-    # Address models
-    "address": Address,
-    "addresses": Address,
-    
-    # Company models
-    "company": CompanyBasic,
-    "companies": CompanyBasic,
-    "company_details": CompanyDetails,
-    "company_fulldetails": CompanyFullDetails,
-    
-    # Contact models
-    "contact": Contact,
-    "contacts": Contact,
-    
-    # Job models
-    "job": Job,
-    "jobs": Job,
-    "job_item": JobItem,
-    "job_task": JobTask,
-    "job_shipment": JobShipment,
-    "job_payment": JobPayment,
-    
-    # User models
-    "user": User,
-    "users": User,
-    
-    # Document models
-    "document": Document,
-    "documents": Document,
-    
-    # Form models
-    "form": Form,
-    "forms": Form,
-    
-    # Lookup models
-    "lookup_value": LookupValue,
-}
-
-
-def get_model_class(resource_name: str) -> Optional[type]:
-    """Get model class for a resource name."""
-    return MODEL_REGISTRY.get(resource_name.lower())
-
-
-def parse_response(data: Any, resource_name: Optional[str] = None) -> Any:
-    """Parse API response into model instances."""
-    if not data:
-        return data
-    
-    model_class = get_model_class(resource_name) if resource_name else None
-    
-    if isinstance(data, list):
-        if model_class and hasattr(model_class, "model_validate"):
-            return [model_class.model_validate(item) for item in data]
-        return data
-    
-    if isinstance(data, dict):
-        if any(key in data for key in ["data", "items", "results"]):
-            return PaginatedResponse.from_dict(data, model_class)
-        
-        if "error" in data:
-            return ErrorResponse.model_validate(data)
-        
-        if model_class and hasattr(model_class, "model_validate"):
-            return model_class.model_validate(data)
-    
-    return data
